+++ permissions/src/epilot/models/shared/orgrole.py
@@ -6,12 +6,8 @@ import dateutil.parser
 from ..shared import grant as shared_grant
 from dataclasses_json import Undefined, dataclass_json
 from datetime import datetime
-from enum import Enum
 from epilot import utils
-from typing import Optional
-
-class OrgRoleType(str, Enum):
-    ORG_ROLE = 'org_role'
+from typing import Any, Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
@@ -29,7 +25,7 @@ class OrgRole:
     r"""Id of an organization"""
     slug: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('slug') }})
     r"""URL-friendly name for the role"""
-    type: OrgRoleType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: Any = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     expires_at: Optional[datetime] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('expires_at'), 'encoder': utils.datetimeisoformat(True), 'decoder': dateutil.parser.isoparse, 'exclude': lambda f: f is None }})
     r"""date and time then the role will expire"""